правки согласно замечаниям
