+++ project/src/components/details/details.tsx
@@ -1,0 +1,43 @@
+import { FC, Fragment } from 'react';
+import { TFilm } from '../../types/film';
+import { getRuntimeString } from '../../utils/time-formatters';
+
+export const Details:FC<{ film: TFilm }> = ({ film }) => {
+  const { director, starring, genre, released, runTime } = film;
+
+  return (
+    <div className="film-card__text film-card__row">
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Director</strong>
+          <span className="film-card__details-value">{director}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Starring</strong>
+          <span className="film-card__details-value">
+            {
+              starring.map((actor) => (<Fragment key={actor}>{actor}, <br/></Fragment>))
+            }
+          </span>
+        </p>
+      </div>
+
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Run Time</strong>
+          <span className="film-card__details-value">
+            {getRuntimeString(runTime)}
+          </span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Genre</strong>
+          <span className="film-card__details-value">{genre}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Released</strong>
+          <span className="film-card__details-value">{released}</span>
+        </p>
+      </div>
+    </div>
+  );
+};
