+++ project/src/const.ts
@@ -17,7 +17,10 @@ export const MAX_REVIEW_TEXT_LENGTH = 400;
 export const PREVIEW_PALY_DELAY_MS = 1000;
 
 export enum MovieTab {
+  Overview = 'Overview',
   Details = 'Details',
-  Reviews = 'Reviews',
-  Overview = 'Overview'
+  Reviews = 'Reviews'
 }
+
+export const MAX_FILMS_PER_PAGE = 8;
+export const MAX_RELATED_MOVIES_LIST_LENGTH = 4;
